requirements.txt
streamlit==1.31.0
streamlit-drawable-canvas==0.9.3
pdfplumber==0.10.3
Pillow==10.2.0